\name{multiply_group}
\alias{multiply_group}
\title{Multiply elements of an environment according to a named vector}
\usage{
  multiply_group(grp, v_weights, input_env, result_env)
}
\arguments{
  \item{grp}{name of the selected group, needs to be part
  of the vector names.}

  \item{v_weights}{named vector}

  \item{input_env}{environment that contains the input}

  \item{result_env}{environment to store the results}
}
\description{
  There's no further description.
}
\author{
  Matthias Bannert
}

